requirements.txt
mpyq==0.2.5
sc2reader==1.8.0
python-dotenv==1.0.1